+++ nodes/Autotask/operations/common/udf-flattening/index.ts
@@ -1,0 +1,3 @@
+import { flattenUdfsOption, addFlattenUdfsOption } from './description';
+
+export { flattenUdfsOption, addFlattenUdfsOption };